backend/requirements.txt
# Web framework
Flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9

# Authentication
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3

# Validation / serialisation
marshmallow==3.21.3

# Async tasks
celery==5.4.0
redis==5.0.7

# HTTP client (AI endpoint)
requests==2.32.3

# Geospatial (optional but recommended)
rasterio==1.3.10
Pillow==10.4.0
netCDF4==1.7.1
h5py==3.11.0

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.2.2
pytest-flask==1.3.0
factory-boy==3.3.1

Flask-Cors
